refactor: 스토리북 arg 추가 및 폴더 위치 수정
